fix offset as off_t (#14)
* fix offset
* bump to 0.8.1

+++ caio/thread_aio.c
@@ -156,7 +156,7 @@ void worker(void *arg) {
     }
 
     int fileno = op->fileno;
-    int offset = op->offset;
+    off_t offset = op->offset;
     int buf_size = op->buf_size;
     char* buf = op->buf;
